Move player related composables into player folder
