+++ src/app/components/basic/WinPlayer.vue
@@ -72,8 +72,8 @@ import { useI18n } from 'vue-i18n'
 import type WinPlayerTime from '@app/components/basic/WinPlayerTime.vue'
 import { usePlayerPlaybackStore } from '@app/stores/playerPlaybackStore.ts'
 import { PlayerState } from '@app/types/types.ts'
-import { useVolumeControl } from '@app/composables/useVolumeControl.ts'
-import { useAudioPlayer } from '@app/composables/useAudioPlayer.ts'
+import { useVolumeControl } from '@app/composables/player/useVolumeControl.ts'
+import { useAudioPlayer } from '@app/composables/player/useAudioPlayer.ts'
 import { useWindows } from '@app/composables/useWindows.ts'
 import { useNowPlayingStatus } from '@app/composables/player/useNowPlayingStatus.ts'
 import { useAuth } from '@app/composables/useAuth.ts'
