+++ src/app/components/windows/WindowPlayer.vue
@@ -38,7 +38,7 @@ import { useI18n } from 'vue-i18n'
 import WinWindow from '@app/components/basic/WinWindow.vue'
 import { isMobile } from '@app/utils/helpers.ts'
 import { Native } from '@mobile/bridge/native.ts'
-import { useVolumeControl } from '@app/composables/useVolumeControl.ts'
+import { useVolumeControl } from '@app/composables/player/useVolumeControl.ts'
 import { useWindows } from '@app/composables/useWindows.ts'
 import { useNowPlayingStatus } from '@app/composables/player/useNowPlayingStatus.ts'
 import { useAuth } from '@app/composables/useAuth.ts'
